+++ NEWS.md
@@ -2,7 +2,9 @@
 
 ## mkin 0.9.41.9000
 
+### Minor changes
 
+- Add plots to `compiled_models` vignette
 
 ## mkin 0.9-41 (2015-11-09)